Fix rewiring issue. RRT star for holonomic seems to works
The problem was that the set of edges from the graph couldn't find edges when the source has been updated due to the rewiring step. Fix was to get all edges as a list instead, and to erase the element at one exact spot. Then the edges from that node, if there was any, got their source updated to be the new node after rewire update